Support for Shelly Dimmer Gen3
